Moved to CssGenerator
